Made TimeLineViewHolder an innerclass of TimeLineAdapter to respect ViewHolder pattern for better clarity. Corrected getMarkerDrawable for better handling nullPointerException
