+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/schedule/TimeLineAdapter.java
@@ -11,7 +11,10 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
+import butterknife.BindView;
+import butterknife.ButterKnife;
 import ch.epfl.sweng.eventmanager.repository.data.Concert;
 import com.github.vipulasri.timelineview.TimelineView;
 
@@ -26,7 +29,7 @@ public class TimeLineAdapter extends RecyclerView.Adapter<TimeLineViewHolder> {
     private Context context;
     private LayoutInflater mLayoutInflater;
 
-    public TimeLineAdapter(List<Concert> feedList) {
+    TimeLineAdapter(List<Concert> feedList) {
         dataList = feedList;
     }
 
@@ -75,7 +78,7 @@ public class TimeLineAdapter extends RecyclerView.Adapter<TimeLineViewHolder> {
 
     @Override
     public int getItemCount() {
-        return (dataList !=null? dataList.size():0);
+        return (dataList != null ? dataList.size() : 0);
     }
 
     private String transformDuration(double duration) {
@@ -84,14 +87,37 @@ public class TimeLineAdapter extends RecyclerView.Adapter<TimeLineViewHolder> {
         return intPart+":"+fractionalPart;
     }
 
+
     private static Drawable getMarkerDrawable(Context context, int drawableResId, int colorFilter) {
         Drawable drawable = VectorDrawableCompat.create(context.getResources(), drawableResId, context.getTheme());
-        try {
+            if (drawable == null) {
+                Log.i("ERROR", "setColorFilter method returned null in TimeLineAdapter while " +
+                        "fetching marker drawable");
+                return null;
+            }
             drawable.setColorFilter(ContextCompat.getColor(context, colorFilter), PorterDuff.Mode.SRC_IN);
-        } catch (NullPointerException n) {
-            Log.i("ERROR", "setColorFilter method returned null in TimeLineAdapter while " +
-                    "fetching marker drawable");
-        }
-        return drawable;
+            return drawable;
+    }
+}
+
+class TimeLineViewHolder extends RecyclerView.ViewHolder {
+
+    @BindView(R.id.text_timeline_date)
+    TextView mDate;
+    @BindView(R.id.text_timeline_artist)
+    TextView mArtist;
+    @BindView(R.id.text_timeline_genre)
+    TextView mGenre;
+    @BindView(R.id.text_timeline_description)
+    TextView mDescription;
+    @BindView(R.id.text_timeline_duration)
+    TextView mDuration;
+    @BindView(R.id.time_marker)
+    TimelineView mTimelineView;
+
+    TimeLineViewHolder(View itemView, int viewType) {
+        super(itemView);
+        ButterKnife.bind(this, itemView);
+        mTimelineView.initLine(viewType);
     }
 }
